feat: transform output name for options
