+++ test/command.test.ts
@@ -10,7 +10,7 @@ test('constructor', () => {
   expect(massarg(opts)).toBeInstanceOf(MassargCommand)
 })
 
-describe('command', () => {
+describe('sub command', () => {
   test('add', () => {
     const command = massarg(opts)
     expect(command.command).toBeInstanceOf(Function)
@@ -33,149 +33,6 @@ describe('command', () => {
         run: jest.fn(),
       }),
     ).toThrow('Expected string, received number')
-  })
-})
-
-describe('option', () => {
-  test('add', () => {
-    const command = massarg(opts)
-    expect(command.option).toBeInstanceOf(Function)
-    expect(
-      command.option({ name: 'test2', description: 'test2', aliases: [], defaultValue: '' }),
-    ).toBeInstanceOf(MassargCommand)
-  })
-  test('validate', () => {
-    expect(() =>
-      massarg(opts).option({
-        name: 'test2',
-        description: 123 as any,
-        aliases: [],
-        defaultValue: '',
-      }),
-    ).toThrow('Expected string, received number')
-  })
-  test('add duplicate', () => {
-    expect(() =>
-      massarg(opts)
-        .option({
-          name: 'test2',
-          description: 'test2',
-          aliases: [],
-          defaultValue: '',
-        })
-        .option({
-          name: 'test2',
-          description: 'test2',
-          aliases: [],
-          defaultValue: '',
-        }),
-    ).toThrow('Option "test2" already exists')
-  })
-  test('add 2 defaults', () => {
-    expect(() =>
-      massarg(opts)
-        .option({
-          name: 'test',
-          description: 'test2',
-          aliases: [],
-          isDefault: true,
-        })
-        .option({
-          name: 'test2',
-          description: 'test2',
-          aliases: [],
-          isDefault: true,
-        }),
-    ).toThrow(
-      'Option "test2" cannot be set as default because option "test" is already set as default',
-    )
-  })
-})
-
-describe('flag', () => {
-  test('add', () => {
-    const command = massarg(opts)
-    expect(command.flag).toBeInstanceOf(Function)
-    expect(command.flag({ name: 'test2', description: 'test2', aliases: [] })).toBeInstanceOf(
-      MassargCommand,
-    )
-  })
-  test('add duplicate', () => {
-    expect(() =>
-      massarg(opts)
-        .flag({ name: 'test2', description: 'test2', aliases: [] })
-        .flag({ name: 'test2', description: 'test2', aliases: [] }),
-    ).toThrow('Flag "test2" already exists')
-  })
-  test('validate', () => {
-    expect(() =>
-      massarg(opts).flag({
-        name: 'test2',
-        description: 123 as any,
-        aliases: [],
-      }),
-    ).toThrow('Expected string, received number')
-  })
-})
-
-describe('example', () => {
-  test('example', () => {
-    const command = massarg(opts)
-    expect(command.example).toBeInstanceOf(Function)
-    expect(command.example({ description: 'test', input: '', output: '' })).toBeInstanceOf(
-      MassargCommand,
-    )
-  })
-})
-
-describe('help', () => {
-  test('default value', () => {
-    const command = massarg(opts)
-    expect(command.helpConfig).toEqual(defaultHelpConfig)
-  })
-
-  test('init', () => {
-    const command = massarg(opts).help({
-      bindOption: true,
-      optionOptions: {
-        namePrefix: '__',
-      },
-    })
-    expect(command.help).toBeInstanceOf(Function)
-    expect(command.helpConfig).toHaveProperty('bindOption', true)
-    expect(command.helpConfig).toHaveProperty('optionOptions.namePrefix', '__')
-    expect(command.helpConfig).toHaveProperty('optionOptions.aliasPrefix', '-')
-    expect(command.helpConfig).toHaveProperty('optionOptions.nameStyle.color', 'yellow')
-  })
-
-  test('binds command', () => {
-    const command = massarg(opts).help({
-      bindCommand: true,
-    })
-    expect(command.help).toBeInstanceOf(Function)
-    expect(command.helpConfig).toHaveProperty('bindCommand', true)
-    expect(command.commands.find((o) => o.name === 'help')).toBeTruthy()
-  })
-
-  test('binds option', () => {
-    const command = massarg(opts).help({
-      bindOption: true,
-    })
-    expect(command.help).toBeInstanceOf(Function)
-    expect(command.helpConfig).toHaveProperty('bindOption', true)
-    expect(command.options.find((o) => o.name === 'help')).toBeTruthy()
-  })
-
-  test('help string', () => {
-    const command = massarg(opts)
-    expect(command.helpString()).toContain(`Usage:`)
-  })
-
-  test('print help', () => {
-    const log = jest.spyOn(console, 'log').mockImplementation(() => {})
-    const command = massarg(opts)
-    command.printHelp()
-    expect(log).toHaveBeenCalled()
   })
 })
